# api/openapi.yaml
openapi: 3.1.0
info:
  title: OneImg API
  version: 1.0.0
  description: Breaking v1 API for OneImg. Legacy /api routes are not supported.
servers:
  - url: /api/v1
...
paths:
...
  /me/passkeys:
    get:
      operationId: listPasskeys
      tags: [Account]
      security: [{ cookieAuth: [] }]
      responses:
        '200': { $ref: '#/components/responses/DataResponse' }
        default: { $ref: '#/components/responses/ProblemResponse' }
...
  /me/passkeys/{id}:
    parameters: [{ $ref: '#/components/parameters/ID' }]
    patch:
      operationId: renamePasskey
      tags: [Account]
      security: [{ cookieAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [name], properties: { name: { type: string } } } } }
      responses:
        '200': { $ref: '#/components/responses/DataResponse' }
        default: { $ref: '#/components/responses/ProblemResponse' }
  /me/passkeys/{id}/revoke:
    parameters: [{ $ref: '#/components/parameters/ID' }]
    post:
      operationId: revokePasskey
      tags: [Account]
      security: [{ cookieAuth: [] }]
      requestBody: { $ref: '#/components/requestBodies/PasswordConfirmation' }
      responses:
        '204': { $ref: '#/components/responses/NoContentResponse' }
        default: { $ref: '#/components/responses/ProblemResponse' }
...
components:
  securitySchemes:
    cookieAuth: { type: apiKey, in: cookie, name: oneimg-session }
    bearerAuth: { type: http, scheme: bearer }
  parameters:
    ID: { name: id, in: path, required: true, schema: { type: integer, minimum: 1 } }
...
  requestBodies:
...
    PasswordConfirmation:
      required: true
      content: { application/json: { schema: { type: object, required: [current_password], properties: { current_password: { type: string, format: password } } } } }
...
  responses:
    ProblemResponse:
      description: RFC 9457 Problem Details
      headers: { X-Request-ID: { schema: { type: string } } }
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
    DataResponse:
      description: Successful response
      headers: { X-Request-ID: { schema: { type: string } } }
      content: { application/json: { schema: { $ref: '#/components/schemas/DataEnvelope' } } }
    NoContentResponse:
      description: Successful response with no content
      headers: { X-Request-ID: { schema: { type: string } } }
...
  schemas:
    DataEnvelope:
      type: object
      required: [data]
      properties: { data: {}, meta: { $ref: '#/components/schemas/Meta' } }
    Meta:
      type: object
      properties:
        pagination:
          type: object
          required: [page, page_size, total, total_pages]
          properties: { page: { type: integer }, page_size: { type: integer }, total: { type: integer, format: int64 }, total_pages: { type: integer, format: int64 } }
    FieldProblem:
      type: object
      required: [field, code, message]
      properties: { field: { type: string }, code: { type: string }, message: { type: string } }
    Problem:
      type: object
      required: [type, title, status, detail, code, instance, request_id]
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        code: { type: string }
        instance: { type: string }
        request_id: { type: string }
        errors: { type: array, items: { $ref: '#/components/schemas/FieldProblem' } }
...
security:
  - cookieAuth: []
  - bearerAuth: []